refactored from types to dataclasses for better expressibility
