+++ ocr_cipher_solver/data_formats/__init__.py
@@ -1,0 +1,1 @@
+from .positional_character import PositionalCharacterSet
