+++ ocr_cipher_solver/data_formats/bounding_box.py
@@ -1,0 +1,11 @@
+import dataclasses
+
+from .position import Position
+
+
+@dataclasses.dataclass
+class BoundingBox:
+    top_left: Position
+    top_right: Position
+    bottom_left: Position
+    bottom_right: Position
