+++ ocr_cipher_solver/data_formats/character.py
@@ -1,0 +1,1 @@
+Character = str
