+++ ocr_cipher_solver/data_formats/position.py
@@ -1,0 +1,8 @@
+import dataclasses
+
+
+@dataclasses.dataclass
+class Position:
+    """Stores x and y position (pixel values)."""
+    x: int
+    y: int
