+++ ocr_cipher_solver/reconstructor/reconstructor.py
@@ -1,6 +1,6 @@
 from PIL import Image
 
-from ocr_cipher_solver.types import PositionalCharacterSet
+from ocr_cipher_solver.data_formats import PositionalCharacterSet
 
 
 ReconstructedImage = Image.Image
